+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 - some GUI elements were not correctly reset after loading or creating a project
+- cancel action of waiting dialogs was not processed if the dialog was closed via pressing escape
 
 ## 1.0.1 - 2015-10-15
 ### Fixed